requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
spacy
pdf2image
pytesseract
Pillow
PyPDF2
pdfplumber
streamlit
google-generativeai
requests
python-dotenv
sentence_transformers
pdfplumber
shap
xgboost
tf-keras
filetype
python-multipart